requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
line-pay==0.1.0
MarkupSafe==1.1.1
python-dotenv==0.10.3
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
tzlocal==2.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
